# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Billing & Inventory Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppUser"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /users/{id}/toggle-active:
    patch:
      operationId: toggleUserActive
      tags: [users]
      summary: Activate or deactivate a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User with updated status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
  /users/{id}/reset-password:
    patch:
      operationId: resetUserPassword
      tags: [users]
      summary: Reset user password (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── Dashboard ─────────────────────────────────────────────────────────
...
    AppUser:
      type: object
      required: [id, name, username, role, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        username:
          type: string
        role:
          type: string
          enum: [admin, user]
        isActive:
          type: boolean
        createdAt:
          type: string
    UserInput:
      type: object
      required: [name, username, password, role]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [admin, user]
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [admin, user]
        isActive:
          type: boolean
        password:
          type: string
          minLength: 6
    ResetPasswordInput:
      type: object
      required: [newPassword]
      properties:
        newPassword:
          type: string
          minLength: 6